chat participant proof of concept
