+++ src/extension.ts
@@ -11,6 +11,7 @@ import { registerMonitoringCommand } from "./services/rotting";
 import { registerPasteMeCommand } from "./copilot-wrapper/paste_me";
 import { activateSidebar } from "./sidebar";
 import { APIManager } from "./services/api_manager";
+import { createCheerleaderChatParticipant } from "./copilot-wrapper/chat_participant";
 
 export function activate(context: vscode.ExtensionContext) {
   // Initialize API Manager first
@@ -30,6 +31,9 @@ export function activate(context: vscode.ExtensionContext) {
   registerMonitoringCommand(context);
   activateEncouragement(context);
   registerPasteMeCommand(context);
+
+  // Register Chat Participant
+  createCheerleaderChatParticipant(context);
 
   // Add disposal of WebSocket service
   context.subscriptions.push(
